Created example API documentation for email account creation & deletion (#2)
* Created example API document for email account addition

Added API template for email account creation.
Saw a post, and started checking. Post URL: https://forum.hestiacp.com/t/how-do-i-create-and-delete-mail-accounts-with-the-api/18328

* Created example API document for email account deletion

Added API template for email account deletion.
Saw a post, and started checking. Post URL: https://forum.hestiacp.com/t/how-do-i-create-and-delete-mail-accounts-with-the-api/18328

* Corrected indentation

NPP caused incorrect indentation earlier

* Corrected indentation

NPP caused incorrect indentation earlier

* Created example API document for email account addition

Added API template for email account creation.
Saw a post, and started checking. Post URL: https://forum.hestiacp.com/t/how-do-i-create-and-delete-mail-accounts-with-the-api/18328

* Created example API document for email account deletion

Added API template for email account deletion.
Saw a post, and started checking. Post URL: https://forum.hestiacp.com/t/how-do-i-create-and-delete-mail-accounts-with-the-api/18328

* Corrected per review request

removed `domain_idn` & `HIDE` which was my oversight.

* Removed domain_idn as suggested

* Updated per review

* Update per review

* Rename add_mail_account.php to add_email_account.php

To be in sync with nodejs and email is correct compared to mail in context

* Rename delete_mail_account.php to delete_email_account.php

To be in sync with nodejs and email is correct compared to mail in context